Refactor: Move evaluator logic to evaluator_v2.py
